+++ pipelines/scripts/check_notebooks_valid.py
@@ -1,4 +1,4 @@
-from pipelines.scripts.util import Util
+from pipelines.scripts.util.util import Util
 import json
 from typing import Dict, Any
 import logging
