+++ pipelines/scripts/deploy_odw_package.py
@@ -1,7 +1,5 @@
+from pipelines.scripts.util.synapse_workspace_manager import SynapseWorkspaceManager
 from pipelines.scripts.config import CONFIG
-from pipelines.scripts.util import Util
-from azure.identity import AzureCliCredential
-import requests
 import argparse
 import json
 from typing import List, Dict, Any
@@ -16,97 +14,6 @@ TARGET_SPARK_POOLS = [
     "pinssynspodwpr",
     "pinssynspodw34"
 ]
-
-
-class SynapseWorkspaceManager():
-    """
-        Class to interact with Synapse Private Link private endpoints
-
-        Note: API/CLI support for this kind of endpoint is quite limited. Only getting endpoints is available
-    """
-    _token = None
-
-    def __init__(self, workspace_name: str, subscription_id: str, resource_group_name: str):
-        self.workspace_name = workspace_name
-        self.ENDPOINT = (
-            f"https://management.azure.com/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/"
-            f"Microsoft.Synapse/workspaces/{workspace_name}"
-        )
-
-    @classmethod
-    def _get_token(cls) -> str:
-        if not (cls._token):
-            cls._token = AzureCliCredential().get_token("https://management.azure.com/.default").token
-        return cls._token
-
-    def get_workspace_packages(self) -> Dict[str, Any]:
-        resp: requests.Response = requests.get(
-            f"{self.ENDPOINT}/libraries?api-version=2021-06-01",
-            headers={"Authorization": f"Bearer {self._get_token()}"}
-        )
-        if "application/json" in resp.headers.get("Content-Type", ""):
-            resp_json = resp.json()
-            if "value" in resp_json:
-                return resp_json["value"]
-            raise ValueError(f"response raised an exception: {json.dumps(resp_json, indent=4)}")
-        raise ValueError(f"http endpoint did not respond with a json object. Received {resp}")
-    
-    def upload_workspace_package(self, package_name: str):
-        return json.loads(
-            Util.run_az_cli_command(
-                [
-                    "az",
-                    "synapse",
-                    "workspace-package",
-                    "upload",
-                    "--workspace-name",
-                    self.workspace_name,
-                    "--package",
-                    package_name,
-                    "--no-progress"
-                ]
-            )
-        )
-    
-    def remove_workspace_package(self, package_name: str):
-        resp = json.loads(
-            Util.run_az_cli_command(
-                [
-                    "az",
-                    "synapse",
-                    "workspace-package",
-                    "delete",
-                    "--workspace-name",
-                    self.workspace_name,
-                    "--package",
-                    package_name,
-                    "--no-wait",
-                    "-y"
-                ]
-            )
-        )
-        # Need to wait for the package to be deleted before continuing
-        return resp
-    
-    def get_spark_pool(self, spark_pool_name: str):
-        resp = requests.Response = requests.get(
-            f"{self.ENDPOINT}/bigDataPools/{spark_pool_name}?api-version=2021-06-01",
-            headers={"Authorization": f"Bearer {self._get_token()}"}
-        )
-        if "application/json" in resp.headers.get("Content-Type", ""):
-            return resp.json()
-        raise ValueError(f"http endpint did not respond with a json object. Received {resp}")
-
-    def update_sparkpool(self, spark_pool_name: str, spark_pool_json: Dict[str, Any]):
-        resp: requests.Response = requests.put(
-            f"{self.ENDPOINT}/bigDataPools/{spark_pool_name}?api-version=2021-06-01",
-            json=spark_pool_json,
-            headers={"Authorization": f"Bearer {self._get_token()}"}
-        )
-        if "application/json" in resp.headers.get("Content-Type", ""):
-            # Need to wait for the spark pool to exit provisioning state
-            return resp.json()
-        raise ValueError(f"http endpint did not respond with a json object. Received {resp}")
 
 
 def get_existing_odw_wheels(workspace_manager: SynapseWorkspaceManager) -> List[Dict[str, Any]]:
@@ -172,7 +79,7 @@ if __name__ == "__main__":
         for spark_pool, spark_pool_json in initial_spark_pool_json_map.items()
     }
     logging.info("Updating spark pool packages")
-    resp = synapse_workspace_manager.update_sparkpool("pinssynspodw34", new_spark_pool_json_map["pinssynspodw34"])
+    resp = synapse_workspace_manager.update_sparkpool("pinssynspodwpr", new_spark_pool_json_map["pinssynspodwpr"])
     logging.info(json.dumps(resp, indent=4))
     logging.info("Removing old workspace packages")
     for package in existing_wheel_names:
