+++ pipelines/scripts/synapse_artifact/synapse_notebook_util.py
@@ -1,6 +1,6 @@
 from pipelines.scripts.synapse_artifact.synapse_artifact_util import SynapseArtifactUtil
 from typing import List, Dict, Any, Set
-from pipelines.scripts.util import Util
+from pipelines.scripts.util.util import Util
 import ast
 from ast2json import ast2json
 import re
